Remove outdated model code and update MM model readers to take out underscore and mm
